requirements.txt
autocommand==2.2.2
inflect==6.0.4
irc==20.1.0
jaraco.collections==4.1.0
jaraco.context==4.3.0
jaraco.functools==3.6.0
jaraco.logging==3.1.2
jaraco.stream==3.0.3
jaraco.text==3.11.1
more-itertools==9.1.0
pydantic==1.10.7
python-dotenv==1.0.0
pytz==2023.3
tempora==5.2.2
typing_extensions==4.5.0
